Extract ReadingScraper lib, update rake task to use it
